refactor: extract migration logic to reusable package
